refactor(layout): deepen artifact placement
